tools/requirements.txt
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
referencing==0.37.0
rpds-py==0.30.0
SQLAlchemy==2.0.46
typing-inspection==0.4.2
typing_extensions==4.15.0
